chinh sua xoa san pham trong gio hang

+++ regal/src/main/java/com/example/regal/controller/GioHangController.java
@@ -85,12 +85,21 @@ public class GioHangController {
     }
 
     // Xóa giỏ hàng theo ID
+//    @DeleteMapping("/delete/{id}")
+//    public ResponseEntity<Void> deleteGioHang(@PathVariable Integer id) {
+//        gioHangService.deleteGioHang(id);
+//        return ResponseEntity.noContent().build();
+//    }
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deleteGioHang(@PathVariable Integer id) {
+        if (!gioHangService.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         gioHangService.deleteGioHang(id);
         return ResponseEntity.noContent().build();
     }
 
+    //
     @PostMapping("/tinh-tong")
     public ResponseEntity<?> kiemTraPhieu(@RequestBody List<Integer> idGioHang) {
         List<GioHang> list = gioHangRepository.findAllById(idGioHang);
